requirements.txt
gradio==3.27.0
Pillow==9.5.0
tqdm==4.65.0
